requirement.txt
absl-py==0.7.1
aip==1.0.0.5
alembic==1.0.10
astor==0.7.1
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
blinker==1.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
easydict==1.9
entrypoints==0.3
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Cache==0.13.1
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gast==0.2.2
grpcio==1.20.0
h5py==2.9.0
idna==2.8
intel-openmp==2019.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
jieba==0.39
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lxml==4.3.3
Mako==1.0.10
marisa-trie==0.7.5
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mkl==2019.0
mock==2.0.0
mysqlclient==1.4.2.post1
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.2
opencv-python==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
passlib==1.7.1
pbr==5.1.3
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
Pygments==2.3.1
PyMySQL==0.9.3
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
python-docx==0.8.10
python-dotenv==0.10.2
python-editor==1.0.4
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
redis==3.2.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
soupsieve==1.9.1
speaklater==1.3
SQLAlchemy==1.3.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
WTForms==2.2.1
